CM-25: move code and name validation to validation classes
